Change default for pylsp.plugins.ruff.executable to null
Change the default value for `pylsp.plugins.ruff.executable` to `null` to
match the default for `python-lsp-ruff`. This fixes #24 and restores the
expected default behaviour of `python-lsp-ruff`, which is to look for
`ruff` in the current Python environment. Without this fix the user has
to either install `ruff` globally, or ensure their editor is running in
a suitably activated virtual environment.

Fixes #24.